graphslam-engine: Move header files into corresponding directories
Organize header files based on their purpose into the following
directories:

NRD => Node Registration Deciders
ERD => Edge Registration Deciders
GSO => GraphSLAM Optimizers
misc => Miscellaneous graph and display window manipulation classes
interfaces => Interfaces for implementing NRDs, ERDs, GSOs etc.
apps_related => Classes that can be used in applications utilizing the
current mrpt-graphslam API
