+++ libs/graphslam/include/mrpt/graphslam/ERD/CICPCriteriaERD_impl.h
