+++ libs/graphslam/include/mrpt/graphslam/ERD/CLoopCloserERD_impl.h
