+++ libs/graphslam/include/mrpt/graphslam/GSO/CLevMarqGSO.h
@@ -27,7 +27,7 @@
 #include <mrpt/utils/TColor.h>
 
 #include <mrpt/graphslam/levmarq.h>
-#include <mrpt/graphslam/CGraphSlamOptimizer.h>
+#include <mrpt/graphslam/interfaces/CGraphSlamOptimizer.h>
 
 #include <iostream>
 #include <string>
