+++ libs/graphslam/include/mrpt/graphslam/NRD/CEmptyNRD.h
@@ -1,4 +1,4 @@
-	/* +---------------------------------------------------------------------------+
+/* +---------------------------------------------------------------------------+
 	 |                     Mobile Robot Programming Toolkit (MRPT)               |
 	 |                          http://www.mrpt.org/                             |
 	 |                                                                           |
@@ -7,63 +7,63 @@
 	 | Released under BSD License. See details in http://www.mrpt.org/License    |
 	 +---------------------------------------------------------------------------+ */
 
-#ifndef CEmptyERD_H
-#define CEmptyERD_H
+#ifndef CEMPTYNRD_H
+#define CEMPTYNRD_H
 
-#include <mrpt/obs/CObservation2DRangeScan.h>
+#include <mrpt/obs/CActionCollection.h>
 #include <mrpt/obs/CSensoryFrame.h>
 #include <mrpt/obs/CRawlog.h>
 
-#include "CEdgeRegistrationDecider.h"
+#include <mrpt/graphslam/interfaces/CNodeRegistrationDecider.h>
 
 namespace mrpt { namespace graphslam { namespace deciders {
 
-/**\brief Empty Edge Registration Decider
+/**\brief Empty Node Registration Decider
  *
  * Handy when you are testing other parts of the application but not the
  * specific registration procedure
+ *
+ * \ingroup mrpt_graphslam_grp
  */
-template<class GRAPH_t=typename mrpt::graphs::CNetworkOfPoses2DInf >
-class CEmptyERD:
-	public mrpt::graphslam::deciders::CEdgeRegistrationDecider<GRAPH_t>
+template<class GRAPH_t=typename mrpt::graphs::CNetworkOfPoses2DInf>
+class CEmptyNRD:
+	public mrpt::graphslam::deciders::CNodeRegistrationDecider<GRAPH_t>
 {
+	typedef typename GRAPH_t::constraint_t::type_value pose_t;
 	public:
-		typedef typename GRAPH_t::constraint_t constraint_t;
+		CEmptyNRD();
+		~CEmptyNRD();
 
-		CEmptyERD();
-		~CEmptyERD();
-
-		bool updateState(
-				mrpt::obs::CActionCollectionPtr action,
+		bool updateState( mrpt::obs::CActionCollectionPtr action,
 				mrpt::obs::CSensoryFramePtr observations,
 				mrpt::obs::CObservationPtr observation );
+		pose_t getCurrentRobotPosEstimation() const;
 
 	private:
-		void registerNewEdge(
-				const mrpt::utils::TNodeID& from,
-				const mrpt::utils::TNodeID& to,
-				const constraint_t& rel_edge );
+		void registerNewNode();
+
 };
 
 //////////////////////////////////////////////////////////////////////////////
 
 template<class GRAPH_t>
-CEmptyERD<GRAPH_t>::CEmptyERD() { }
+CEmptyNRD<GRAPH_t>::CEmptyNRD() { }
 template<class GRAPH_t>
-CEmptyERD<GRAPH_t>::~CEmptyERD() { }
+CEmptyNRD<GRAPH_t>::~CEmptyNRD() { }
 
 template<class GRAPH_t>
-bool CEmptyERD<GRAPH_t>::updateState(
+bool CEmptyNRD<GRAPH_t>::updateState(
 		mrpt::obs::CActionCollectionPtr action,
 		mrpt::obs::CSensoryFramePtr observations,
-		mrpt::obs::CObservationPtr observation ) {return true;}
+		mrpt::obs::CObservationPtr observation )  {return false;}
 
 template<class GRAPH_t>
-void CEmptyERD<GRAPH_t>::registerNewEdge(
-		const mrpt::utils::TNodeID& from,
-		const mrpt::utils::TNodeID& to,
-		const constraint_t& rel_edge ) { }
+void CEmptyNRD<GRAPH_t>::registerNewNode() { }
+
+template<class GRAPH_t>
+typename GRAPH_t::constraint_t::type_value
+CEmptyNRD<GRAPH_t>::getCurrentRobotPosEstimation() const {return pose_t();}
 
 } } } // end of namespaces
 
-#endif /* end of include guard: CEmptyERD_H */
+#endif /* end of include guard: CEMPTYNRD_H */
