+++ libs/graphslam/include/mrpt/graphslam/NRD/CFixedIntervalsNRD_impl.h
