+++ libs/graphslam/include/mrpt/graphslam/NRD/CICPCriteriaNRD_impl.h
