+++ libs/graphslam/include/mrpt/graphslam/misc/CRangeScanRegistrationDecider.h
