+++ libs/graphslam/include/mrpt/graphslam/misc/CRangeScanRegistrationDecider_impl.h
