+++ libs/graphslam/include/mrpt/graphslam/misc/CWindowObserver.h
@@ -10,13 +10,14 @@
 #ifndef CWINDOWOBSERVER_H
 #define CWINDOWOBSERVER_H
 
-#include "link_pragmas.h"
 #include <mrpt/gui/CBaseGUIWindow.h>
 #include <mrpt/utils/CTicTac.h>
 #include <mrpt/utils/TParameters.h>
 #include <mrpt/utils/CObserver.h>
 #include <mrpt/opengl/gl_utils.h>
 #include <mrpt/system/string_utils.h>
+
+#include <mrpt/graphslam/link_pragmas.h>
 
 #include <iostream>
 #include <string>
@@ -44,12 +45,12 @@ public:
 	/**\brief Make new keystrokes available in the help message box.
 	 *
 	 * Classes with access to the CWindowObserver can use this method to add
-	 * keystrokes according to their needs 
+	 * keystrokes according to their needs
 	 */
 	void registerKeystroke(const std::string key_str, const std::string key_desc);
 
 protected:
-	void OnEvent(const mrpt::utils::mrptEvent &e); 
+	void OnEvent(const mrpt::utils::mrptEvent &e);
 
 private:
 	bool m_showing_help, m_hiding_help;
@@ -58,7 +59,7 @@ private:
 
 	/**\brief Map from registered char_code (std::string to support <C-c>) to boolean
 	 * true/false indicating whether it has been pressed since previous time
-	 * checked 
+	 * checked
 	 */
 	std::map<std::string, bool> m_key_codes_to_pressed;
 	mrpt::utils::CTicTac  m_tim_show_start, m_tim_show_end;
