+++ libs/graphslam/src/TGraphSlamResources.cpp
@@ -7,7 +7,7 @@
    | Released under BSD License. See details in http://www.mrpt.org/License    |
    +---------------------------------------------------------------------------+ */
 
-#include <mrpt/graphslam/TGraphSlamResources.h>
+#include <mrpt/graphslam/apps_related/TGraphSlamResources.h>
 
 namespace mrpt { namespace graphslam { namespace supplementary {
