+++ share/mrpt/config_files/graphslam-engine/laser_odometry.ini
@@ -45,7 +45,7 @@ registration_max_angle = 8 // degrees
 ; number of ICP goodness values to consider for estimating the current ICP threshold
 sliding_win_size = 3
 ; options:  mean, median
-class_verbosity = 1
+class_verbosity = 0
 
 
 ########################################################
@@ -57,7 +57,7 @@ ICP_goodness_thresh = 0.80 // threshold for accepting the ICP constraint in the 
 ; Either check ICP using distance OR node count from current node
 ICP_max_distance =  5 // maximum distance for checking other nodes for ICP constraints
 ;ICP_max_distance =  -1 // check against all other nodes
-class_verbosity = 1
+class_verbosity = 0
 
 
 [OptimizerParameters]
