+++ share/mrpt/config_files/graphslam-engine/odometry_2DRangeScans.ini
@@ -36,7 +36,7 @@ ground_truth_file_format = NavSimul // variable for determining how to parse the
 ; LVL_INFO  => 1,
 ; LVL_WARN  => 2,
 ; LVL_ERROR => 3
-class_verbosity = 1
+class_verbosity = 0
 
 
 ; min node difference for an edge to be considered as a loop closure
@@ -51,7 +51,7 @@ LC_min_nodeid_diff= 50
 registration_max_distance = 0.2 // meters
 ;registration_max_distance = 1 // meters
 registration_max_angle = 30 // degrees
-class_verbosity = 1
+class_verbosity = 0
 
 
 ########################################################
