+++ share/mrpt/config_files/graphslam-engine/ros_laser_odometry.ini
@@ -41,30 +41,30 @@ class_verbosity = 1
 [NodeRegistrationDeciderParameters]
 
 registration_max_distance = 0.1 // meters
-registration_max_angle = 8 // degrees
+registration_max_angle = 10 // degrees
 
 ; number of ICP goodness values to consider for estimating the current ICP threshold
 sliding_win_size = 3
 ; options:  mean, median
-class_verbosity = 1
+class_verbosity = 0
 
 
 ########################################################
 [EdgeRegistrationDeciderParameters]
 
 ; Works on the 2D observation-only dataset
-ICP_goodness_thresh = 0.80 // threshold for accepting the ICP constraint in the graph
+ICP_goodness_thresh = 0.90 // threshold for accepting the ICP constraint in the graph
 
 ; Either check ICP using distance OR node count from current node
-ICP_max_distance =  5 // maximum distance for checking other nodes for ICP constraints
+ICP_max_distance =  10 // maximum distance for checking other nodes for ICP constraints
 ;ICP_max_distance =  -1 // check against all other nodes
-class_verbosity = 1
+class_verbosity = 0
 
 
 [OptimizerParameters]
 
 optimization_on_second_thread = false
-optimization_distance = 3;
+optimization_distance = 1;
 
 ; Levenberg-Marquardt parameters
 verbose = false
